Event/requirements.txt
Flask==0.12
Flask-Cors==3.0.2
Flask-RESTful==0.3.6
Jinja2==2.9.4
requests==2.12.4
SQLAlchemy==1.1.5
kafka-python==1.3.3
mysql-python==1.2.5
